implement proof gen data submitter
