+++ core/node/proof_data_handler/Cargo.toml
@@ -24,6 +24,7 @@ axum.workspace = true
 tokio.workspace = true
 tower-http = { workspace = true, features = ["compression-zstd", "decompression-zstd"] }
 tracing.workspace = true
+reqwest.workspace = true
 
 [dev-dependencies]
 hyper.workspace = true
