+++ zkstack_cli/crates/config/src/chain.rs
@@ -207,3 +207,4 @@ impl FileConfigWithDefaultName for ChainConfigInternal {
 }
 
 impl ZkStackConfig for ChainConfigInternal {}
+// us-docker.pkg.dev/matterlabs-infra/matterlabs-docker/proof-fri-gpu-compressor-gar:2.0-0.26.0-62aea8b-1737712922953
